// ui/src/app/api/models/account-type.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Account type. The first digit of the account code is derived from it:
 * ASSET=1000, LIABILITY=2000, EQUITY=3000, INCOME=4000, EXPENSE=5000.
 */
export type AccountType = 'ASSET' | 'LIABILITY' | 'EQUITY' | 'INCOME' | 'EXPENSE';